Lazily init DNS cache
